Separate files for testing and implement hydration week view functionality
